neurotorch/requirements.txt
numpy>=2.1.2
pandas>=2.2.3
scipy>=1.14.1
matplotlib>=3.9.2
scikit-image>=0.24.0
xarray>=2024.9.0
pims>=0.7
pyimagej==1.5.0
psutil
fsspec>=2024.9.0
requests>=2.32.3
#trace_selector>=0.2.14

#Deprecated

#Pillow
#fsspec
#Requests